reviwes redsighn

+++ app/src/main/java/com/example/myapplication2/model/Comment.java
@@ -53,7 +53,7 @@ public class Comment {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(float rating) {
         this.rating = rating;
     }
 }
